created abstract controller subscriber and combined some modules
