--- machine_common_sense/controller_event_payload.py
@@ -1,4 +1,0 @@
-class ControllerEventPayload():
-
-    def __init__(self):
-        self.step_number = 0
